backend/requirements.txt
openai
python-dotenv
fastapi
uvicorn[standard]
vercel
langfuse
pymongo[srv]
passlib
email-validator
authlib
